Refactor: extract `CreateSiteState` out of preview vm to help splitting
